chore(api): split post-apply audit emission into sibling file
